Added Server Teleporter Module + Fixed Player Pickup & More!
I decided today that maybe we should move the server teleporter from the
autorun folder on the sandbox server and maybe make it an official
module, so after re-organizing the mess I did so. I also fixed the
pickup code which I guess was causing all the player pickup issues. As
for the christmas module, I organized the entity variables better.