[Major Update] Converted to addon + more
I finally decided to do some fixing so I converted the entire system to
an addon to allow us to stay more organized with entities + assets. I
also fixed the teleporter code converting the code to an actural SENT
instead of just using a prop_physics which is very bad practice.